Do some refactoring.
