+++ pyproject.toml
@@ -1,13 +1,3 @@
-[project]
-name = 'api_authorisation_token_scraper'
-version = "0.0.1"
-description = 'Scrape authorisation tokens for various APIs.'
-requires-python = ">=3.9"
-authors = [
-    { email = 'bencapadbury+github@protonmail.com' },
-    { name = 'Ben Padbury' },
-]
-
 [build-system]
 requires = [
     "setuptools",
