+++ setup.cfg
@@ -1,5 +1,5 @@
 [metadata]
-name = api_authorisation_token_scraper
+name = api-authorisation-token-scraper
 version = 0.0.1
 author = Ben Padbury
 author_email = bencapadbury+github@protonmail.com
