+++ src/authorisation_token/api/apple/podcasts.py
@@ -1,4 +1,4 @@
-from api_authorisation_token.scraper import Scraper
+from authorisation_token.scraper import Scraper
 
 
 class Podcasts(Scraper):
