+++ tests/authorisation_token/api/apple/music_test.py
@@ -1,7 +1,7 @@
-from api_authorisation_token.api.apple.podcasts import Podcasts
+from authorisation_token.api import Music
 
 # Get an instance of the system we will be testing.
-systemUnderTest = Podcasts()
+systemUnderTest = Music()
 
 
 # Test that we get back an actual bearer token.
